Art Added
Added the sprites I've made so far such as the player, flower and grass. Added a post processing effect on the camera which I can always change the intensity of the effects. Froze the players z rotation so the sprite will stay upright. Needs to be animated and player and flower collision boxes may need to shrink to fit the sprites.